small changes of units in celestial bodies

+++ Mercury/Planet.cpp
@@ -1,6 +1,6 @@
 #include "Planet.h"
 
-Planet::Planet(int orbit, int mass, int radius)
+Planet::Planet(int orbit, uint64_t mass, int radius)
 	:CelestialBody(orbit, radius, mass)
 {
 	type = CelestialBodyType::Planet;
