+++ Mercury/Star.cpp
@@ -1,6 +1,6 @@
 #include "Star.h"
 
-Star::Star(int orbit, int radius, int mass, float luminosity, float metallicity)
+Star::Star(int orbit, int radius, uint64_t mass, float luminosity, float metallicity)
 	:CelestialBody(orbit, radius, mass),
 	luminosity(luminosity), metallicity(metallicity)
 {
